+++ packages/sdk/src/nft/mint/domain.ts
@@ -1,5 +1,5 @@
 import { ItemId } from "@rarible/api-client"
-import { IBlockchainTransaction } from "@rarible/sdk-transaction/src/domain"
+import { IBlockchainTransaction } from "@rarible/sdk-transaction"
 import { AbstractPrepareResponse } from "../../common/domain"
 import { MintRequest } from "./mint-request.type"
 import { PrepareMintRequest } from "./prepare-mint-request.type"
@@ -11,6 +11,9 @@ type MintResponseCommon = {
 	itemId: ItemId
 }
 
+/**
+ * Mint type
+ */
 export enum MintType {
 	OFF_CHAIN = "off-chain",
 	ON_CHAIN = "on-chain"