+++ packages/transaction/src/index.ts
@@ -1,5 +1,6 @@
 import { Blockchain } from "@rarible/api-client"
 import type { EthereumTransaction } from "@rarible/ethereum-provider"
+import { OperationResult } from "tezos-sdk-module/dist/common/base"
 import { IBlockchainTransaction } from "./domain"
 
 export class BlockchainEthereumTransaction implements IBlockchainTransaction {
@@ -15,3 +16,19 @@ export class BlockchainEthereumTransaction implements IBlockchainTransaction {
 		}
 	}
 }
+
+export class BlockchainTezosTransaction implements IBlockchainTransaction {
+	blockchain: Blockchain = "TEZOS"
+	constructor(public transaction: OperationResult) {}
+
+	async wait() {
+		await this.transaction.confirmation()
+
+		return {
+			blockchain: this.blockchain,
+			hash: this.transaction.hash,
+		}
+	}
+}
+
+export * from "./domain"