+++ CHANGELOG.md
@@ -3,6 +3,111 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.11](https://github.com/rarible/sdk/compare/v0.1.10...v0.1.11) (2021-10-17)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.1.10](https://github.com/rarible/sdk/compare/v0.1.9...v0.1.10) (2021-10-17)
+
+
+### Bug Fixes
+
+* minor fill change ([a924614](https://github.com/rarible/sdk/commit/a9246142c05a8a65de9e5b34733e8c5561fe1a6d))
+
+
+
+
+
+## [0.1.9](https://github.com/rarible/sdk/compare/v0.1.8...v0.1.9) (2021-10-16)
+
+
+### Bug Fixes
+
+* use tezos-sdk-module ([184304f](https://github.com/rarible/sdk/commit/184304f2e58ade0dae1ac16a814f6757949ab790))
+
+
+
+
+
+## [0.1.8](https://github.com/rarible/sdk/compare/v0.1.7...v0.1.8) (2021-10-16)
+
+
+### Bug Fixes
+
+* simplify createRaribleSdk function ([e8f6b06](https://github.com/rarible/sdk/commit/e8f6b0676f9712573f29af31ad7e10cc79964ac5))
+
+
+
+
+
+## [0.1.7](https://github.com/rarible/sdk/compare/v0.1.6...v0.1.7) (2021-10-16)
+
+
+### Bug Fixes
+
+* files section in package.json ([119665b](https://github.com/rarible/sdk/commit/119665b8ce3c7b2ff99c23b456a3c9e8bab28b5f))
+
+
+
+
+
+## [0.1.6](https://github.com/rarible/sdk/compare/v0.1.5...v0.1.6) (2021-10-16)
+
+
+### Bug Fixes
+
+* minor change ([3c06b84](https://github.com/rarible/sdk/commit/3c06b84efd1a1e46cb3b044453a6455a0b40e156))
+
+
+
+
+
+## [0.1.5](https://github.com/rarible/sdk/compare/v0.1.4...v0.1.5) (2021-10-16)
+
+
+### Bug Fixes
+
+* imports ([e0b45b4](https://github.com/rarible/sdk/commit/e0b45b4766f26b4fe01556b41c72e187b2e60253))
+
+
+
+
+
+## [0.1.4](https://github.com/rarible/sdk/compare/v0.1.3...v0.1.4) (2021-10-16)
+
+
+### Bug Fixes
+
+* support for sdk and client updates ([26b59ac](https://github.com/rarible/sdk/commit/26b59aca203ef61c08380077b1d67a675303985a))
+* tezos sell ([339b328](https://github.com/rarible/sdk/commit/339b328c998844b089a70fbad345aa17cce7a5ae))
+
+
+
+
+
+## [0.1.3](https://github.com/rarible/sdk/compare/v0.1.2...v0.1.3) (2021-10-15)
+
+
+### Bug Fixes
+
+* bump ([d8dc39a](https://github.com/rarible/sdk/commit/d8dc39a41c3626954ad24284faa5b8d3854d46b4))
+
+
+
+
+
+## [0.1.2](https://github.com/rarible/sdk/compare/v0.1.1...v0.1.2) (2021-10-15)
+
+**Note:** Version bump only for package root
+
+
+
+
+
 ## [0.1.1](https://github.com/rarible/sdk/compare/v0.1.0...v0.1.1) (2021-10-14)